+++ src/pages/filter/dtree/components/query-builder/ui/trace-variant/trace-variant-results.tsx
